Revert "ERRORS WITH SAVING LINEAR"
This reverts commit f746e06689035395162bb5360ab2c65e329b851f.

+++ lib/domain/model/linear_model.dart
@@ -7,7 +7,7 @@ import 'package:to_do/data/drift/drift_db.dart';
 class LinearModel {
   final int id;
   final int noteId;
-  final double strokeWidth;
+  final int strokeWidth;
   final String colorHex;
   final List<double> dxPositions;
   final List<double> dyPositions;
